Move query string hooks from admin to nav
