+++ client/src/hooks/admin.ts
@@ -1,6 +1,5 @@
 import React from 'react';
 import memoize from 'lodash/memoize';
-import { useLocation } from 'react-router-dom';
 // const memoize = (func: Function) => func;
 
 /**
@@ -188,22 +187,4 @@ export function useCamperLookup(eventId: CtxId): CamperLookup {
   const result = createCamperLookup(registrations, campers, eventId);
 
   return result;
-}
-
-// A custom hook that builds on useLocation to parse
-// the query string for you.
-export function useQuery(type: undefined): URLSearchParams;
-export function useQuery(type: string): string;
-export function useQuery(name?: string) {
-  const params = new URLSearchParams(useLocation().search);
-
-  if (name) return params.get(name);
-
-  return params;
-}
-
-export function useQueryLookup() {
-  const searchParams = new URLSearchParams(useLocation().search);
-
-  return Object.fromEntries(searchParams.entries());
 }
